macro_detector_shared/requirements.txt
Flask
openai
python-dotenv
reportlab
requests
